[refactor] auto login 코드 수정
