+++ client/src/components/Ingredients/TextAreaItem/useTextAreaItem.ts
@@ -42,7 +42,7 @@ const useTextAreaItem = () => {
 
   const debouncedRecommendLatex = useDebounce<[string]>(recommendLatex, 200);
 
-  const onChangeHandler = (size: 'mini' | 'big') => (e: React.ChangeEvent<HTMLTextAreaElement>) => {
+  const onChangeHandler = (size: string) => (e: React.ChangeEvent<HTMLTextAreaElement>) => {
     if (size === 'big' && isShow) {
       debouncedRecommendLatex(e.target.value);
     }
