+++ client/src/components/Meal/MenuBar/style.ts
@@ -9,26 +9,6 @@ export const MenuBarContainer = styled.div`
   justify-content: space-around;
 `;
 
-export const LabelWrapper = styled.div`
-  display: flex;
-  flex-direction: column;
-  width: 70px;
-  min-width: 50px;
-  padding: 10px;
-  text-align: center;
-  font-weight: bold;
-  border: 1px solid #d4d4d5;
-  border-right: none;
-  height: 100%;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-`;
-
-export const LabelIcon = styled.div``;
-
-export const LabelText = styled.div``;
-
 export const OutputWrapper = styled.div`
   width: 30%;
   margin: 10px;
@@ -57,12 +37,6 @@ export const IconWrapper = styled.div`
   & i {
     width: 100% !important;
   }
-`;
-
-export const TextAreaItemWrapper = styled.div`
-  width: 30%;
-  margin: 10px;
-  display: flex;
 `;
 
 export const Logo = styled.div`
