+++ client/src/components/Meal/MenuBarOutput/style.ts
@@ -1,0 +1,27 @@
+import styled from '@emotion/styled';
+
+export const OutputWrapper = styled.div`
+  width: 30%;
+  margin: 10px;
+  display: flex;
+`;
+
+export const LabelWrapper = styled.div`
+  display: flex;
+  flex-direction: column;
+  width: 70px;
+  min-width: 50px;
+  padding: 10px;
+  text-align: center;
+  font-weight: bold;
+  border: 1px solid #d4d4d5;
+  border-right: none;
+  height: 100%;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+`;
+
+export const LabelIcon = styled.div``;
+
+export const LabelText = styled.div``;
