+++ client/src/lib/constants/constants.ts
@@ -1,4 +1,23 @@
+export const INPUT_ICON = '➕';
+
+export const INPUT_TEXT = '입력';
+
+export const OUTPUT_ICON = '✖️';
+
+export const OUTPUT_TEXT = '출력';
+
+export const ICON_SIZE = {
+  mini: 'mini',
+  tiny: 'tiny',
+  small: 'small',
+  large: 'large',
+  big: 'big',
+  huge: 'huge',
+  massive: 'massive',
+};
+
 export const TAB_LIMIT = 4;
+
 export const FONT_SIZE_LISTS = [
   { size: '10', checked: false },
   { size: '15', checked: true },
@@ -6,8 +25,11 @@ export const FONT_SIZE_LISTS = [
   { size: '25', checked: false },
   { size: '30', checked: false },
 ];
+
 export const NEED_LOGIN_ICON = 'dont';
+
 export const NO_LIST_ICON = 'warning';
+
 export enum AlertMessage {
   NEED_LOGIN_MESSAGE = '로그인이 필요합니다.',
   NO_LIST_MESSAGE = '항목이 없습니다.',
