+++ client/src/lib/utils/util.ts
@@ -1,9 +1,8 @@
 import { Action } from 'redux';
 import React from 'react';
-import { API } from '@apis/common';
 import { getToken } from './token';
+import { API } from '@apis/common';
 import { userLogin } from '@contexts/user';
-import { useDispatch } from 'react-redux';
 
 export function createAction<T, P>(type: T): (payload: P) => Action<T> & { payload: P };
 export function createAction<T>(type: T): () => Action<T>;
@@ -20,20 +19,6 @@ export function getImageURL(imageUrl: string | undefined) {
   return process.env.NODE_ENV === 'development'
     ? imageUrl && `./image/${imageUrl}`
     : imageUrl && chrome.extension.getURL(`image/${imageUrl}`);
-}
-
-export async function checkLogin() {
-  const dispatch = useDispatch();
-  const token = await getToken();
-  if (!token) return;
-  const response = await API.post('/auth/autologin', '', {
-    headers: {
-      Authorization: token,
-    },
-  });
-
-  const { userId } = response.data.results;
-  dispatch(userLogin(userId));
 }
 
 function observe() {
